Update filename in script links
The commit corrects the filenames used in script links in the README.md file. Originally, the file was referred to as "tickwatch.min.js"; now, it has been correctly updated to "TickWatch.min.js". This change is important as filenames in scripts are case sensitive and this could potentially solve loading issues.