Update TickWatch-js to version 1.0.2 and refine imports
Version for TickWatch-js has been updated to 1.0.2 throughout the code base, reflected in the banner messages and the package manifests. Also, the import paths in various JavaScript files have been adjusted to enhance the organization of the file structure. Notably, the usage of an alias object in vite.config.js has been removed.