Add 'displayOnly' mode to TickWatch.js
In this update, we introduced the 'displayOnly' mode to TickWatch.js. This mode supports displaying numeric data on the digital clock interface without a timer. The size of the display is customizable via the 'displaySize' parameter. Associated functions and style adjustments were made to ensure smooth operation of this feature.